changed styling and bio location
