+++ src/javascripts/components/projectPhotos/projectPhotos.js
@@ -1,0 +1,22 @@
+import productCardPhoto from '../../../screenshots/productCards.png';
+import PetAdoption from '../../../screenshots/PetAdoption2.png';
+import BearWatcher from '../../../screenshots/'
+
+
+cont projectPhotos = [
+  {
+    name: "Product Cards",
+    image: `${productCardPhoto}`,
+  },
+  {
+    name: "Pet Adoption",
+    image: `${PetAdoption}`,
+  }
+{
+  name: "Bear Watcher",
+    image: `${PetAdoption}`,
+}
+
+];
+
+export default { projectPhotos };
